Move ParticipantBuffer from the entities package to the usecses package.
